Count Zoom Chat Message

This function counts how many times each participants replies in the Zoom chat box and also group messages entries per participants.

Usage

zoom_chat_count(
  data,
  collapse = " ~ ",
  extract_id = FALSE,
  id_regex = "[:digit:]+"
)

Arguments

data

a tibble with class "zoom_chat".

collapse

(Character) deliminator character to separate each entries of participants

extract_id

(Character) Whether to extract ID number from Name column or not.

id_regex

(Character) If extract_id = TRUE, This would specify regular expression ({stringr}'s style) to extract ID.

Value

A tibble with the following columns:

ID: If extract_id = TRUE, this is the ID of the participants.

Name: Name of each participants

Message_Count: Counts of how many times each participants replies

Content: All replies of each participants, separated by collapse argument.